challenges/Misc/签到/src/requirements.txt
flask==3.0.3
Werkzeug==3.0.3
Pillow==10.4.0
requests==2.32.3

